+++ app/src/main/java/com/fieldbook/tracker/database/dao/ObservationDao.kt
@@ -21,6 +21,16 @@ class ObservationDao {
 
     companion object {
 
+        fun getById(id: String): ObservationModel? = withDatabase { db ->
+
+            db.query(Observation.tableName,
+                where = "${Observation.PK} = ?",
+                whereArgs = arrayOf(id))
+                .toFirst()
+                .let { ObservationModel(it) }
+
+        }
+
         fun getAll(): Array<ObservationModel> = withDatabase { db ->
 
             db.query(Observation.tableName)
@@ -611,6 +621,19 @@ class ObservationDao {
                 "internal_id_observation = ?",
                 arrayOf(observation.internal_id_observation.toString())
                 )
+        }
+
+        fun updateObservationValue(
+            id: Int,
+            value: String
+        ) = withDatabase { db ->
+            db.update(Observation.tableName,
+                ContentValues().apply {
+                    put("value", value)
+                },
+                "${Observation.PK} = ?",
+                arrayOf(id.toString())
+            )
         }
 
         //TODO